refactor: bonus_balance to integer type

+++ database/migrations/2025_10_29_133347_add_bonus_balance_to_users_table.php
@@ -13,7 +13,7 @@ return new class () extends Migration {
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table): void {
-            $table->bigInteger('bonus_balance')->default(0)->nullable();
+            $table->integer('bonus_balance')->default(0)->nullable();
         });
     }
